🔄 refactor: home template. Context: Sections are now in external file.  Changes to be committed: 	new file:   src/sections/home/index.jsx 	modified:   src/templates/home.jsx 	new file:   utils/index.ts
