+++ utils/index.ts
@@ -1,0 +1,9 @@
+const linkedin = 'https://linkedin.com/in/gabrielbencomo';
+const twitter = 'https://twitter.com/gabrielbca15';
+const github = 'https://github.com/gabrielba15';
+
+export const social = {
+	linkedin,
+	twitter,
+	github,
+};
